requirements.txt
pyTelegramBotAPI
alembic
python-dotenv
sqlalchemy
